pi-qibla
The Stand-Alone Qibla Direction Finder is a portable device enabling a user to determine the direction of the Qibla, or the proper orientation to face the Kaaba in Mecca, Saudi Arabia. The device contains a GPS receiver, compass module and LCD display, controlled by a single-board computer. The Stand-Alone gadget finds it’s position on Earth by the GPS receiver; the Qibla direction is determined by the GPS coordinates and a spherical triangle equation, resulting in an angle in degrees from true north. The magnetic compass sensor orients the device and magnetic declination is added to adjust the compass heading from magnetic north to true north. An LCD screen is provided as a visual indicator aligning the user in the correct direction.
In this project, we have built a programmable platform specifically used for finding the direction of the Qibla. This scaleable platform can be reprogrammed to find the direction to any point of interest. With just a slight change of software, this prototype could be modified for use in multiple applications.